add destination to standard destination test names
