+++ airbyte-integrations/connectors/destination-meilisearch/src/test-integration/java/io/airbyte/integrations/destination/meilisearch/MeiliSearchDestinationStandardTest.java
@@ -42,7 +42,7 @@ import java.util.stream.Collectors;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.utility.DockerImageName;
 
-public class MeiliSearchStandardTest extends DestinationStandardTest {
+public class MeiliSearchDestinationStandardTest extends DestinationStandardTest {
 
   private static final Integer DEFAULT_MEILI_SEARCH_PORT = 7700;
   private static final Integer EXPOSED_PORT = 7701;
