+++ airbyte-integrations/connectors/destination-mssql/src/test-integration/java/io/airbyte/integrations/destination/mssql/MSSQLDestinationStandardTestSSL.java
@@ -43,8 +43,9 @@ import org.jooq.JSONFormat.RecordFormat;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.containers.MSSQLServerContainer;
+import org.testcontainers.utility.DockerImageName;
 
-public class MSSQLStandardTest extends DestinationStandardTest {
+public class MSSQLDestinationStandardTestSSL extends DestinationStandardTest {
 
   private static final JSONFormat JSON_FORMAT = new JSONFormat().recordFormat(RecordFormat.OBJECT);
 
@@ -65,6 +66,7 @@ public class MSSQLStandardTest extends DestinationStandardTest {
         .put("username", db.getUsername())
         .put("password", db.getPassword())
         .put("schema", "testSchema")
+        .put("ssl_method", "encrypted_trust_server_certificate")
         .build());
   }
 
@@ -107,6 +109,12 @@ public class MSSQLStandardTest extends DestinationStandardTest {
   protected List<JsonNode> retrieveNormalizedRecords(TestDestinationEnv env, String streamName, String namespace)
       throws Exception {
     String tableName = namingResolver.getIdentifier(streamName);
+    // Temporarily disabling the behavior of the ExtendedNameTransformer, see (issue #1785) so we don't
+    // use quoted names
+    // if (!tableName.startsWith("\"")) {
+    // // Currently, Normalization always quote tables identifiers
+    // //tableName = "\"" + tableName + "\"";
+    // }
     return retrieveRecordsFromTable(tableName, namespace);
   }
 
@@ -139,7 +147,8 @@ public class MSSQLStandardTest extends DestinationStandardTest {
 
   @BeforeAll
   protected static void init() {
-    db = new MSSQLServerContainer<>("mcr.microsoft.com/mssql/server:2019-latest").acceptLicense();
+    db = new MSSQLServerContainer<>(DockerImageName.parse("airbyte/mssql_ssltest:dev").asCompatibleSubstituteFor("mcr.microsoft.com/mssql/server"))
+        .acceptLicense();
     db.start();
   }
